New wording of license exemptions
I wrote a new version of the exception. In practicallity, they do the
exact same thing but you can't really take a gamble when it comes to
legal stuff. We made everything super clear. And yes, i can do this
because no other third party has submitted code yet. All of it is under
my copyright.

Signed-off-by: Joseph Umana <joseph@josephumana.com>