Add `free_s` and `memset_s` functions
`free_s` is a function that frees the page while also zeroing it.
`memset_s` is of course a dependency for this which is `memset` without
compiler optimizations. `free_s` and `memset_s` should be used for
sensitive data like crypto keys and passwords to prevent attacks where
memory is freed and allocated in order to read the reminents of the page
that may contain sensitive data.

Signed-off-by: Joseph Umana <joseph@josephumana.com>